api-proxy/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.27.0
psycopg2-binary>=2.9.9
